Moved FirePopup files out of temp
