+++ src/features/common/FirePopup/FirePopup.module.scss
@@ -1,4 +1,4 @@
-@import '../../theme/theme';
+@import '../../../theme/theme';
 
 .arrow {
   display: block;
