+++ src/temp/stories/projectDetails/FirePopup.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Popup from '../../FirePopup';
+import Popup from '../../../features/common/FirePopup';
 
 const meta: Meta<typeof Popup> = {
   title: 'Projects/Details/Popup',
